requirements.txt
numpy==2.3.1
opencv_python==4.10.0.84
Pillow==10.1.0
Pillow==11.3.0
scikit_learn==1.7.0
